+++ soccertime/settings.py
@@ -264,3 +264,4 @@ SILENCED_SYSTEM_CHECKS = [
     "security.W005",
     "security.W021",
 ]
+